requirements.txt
python-telegram-bot==21.4
python-dotenv==1.0.1
SQLAlchemy==2.0.36
fastapi==0.115.6
uvicorn==0.32.1
tzdata==2024.2
psycopg[binary]==3.2.3

